test(codex): split CodeGraph node support cases
