+++ .omo/evidence/20260620-codegraph-bootstrap-node26/focused-component-checks-after-test-split.txt
@@ -1,0 +1,59 @@
+$ cd packages/omo-codex/plugin/components/codegraph && bun run typecheck
+$ tsc --noEmit
+
+$ cd packages/omo-codex/plugin/components/codegraph && bun run test
+$ bun test test/*.test.ts
+bun test v1.3.14 (0d9b296a)
+
+test/provisioned-node-guard.test.ts:
+(pass) CodeGraph provisioned launcher Node guard > #given provisioned CodeGraph binary #when serve runs under unsupported local Node #then it trusts the launcher [2.44ms]
+(pass) CodeGraph provisioned launcher Node guard > #given provisioned CodeGraph exists #when SessionStart worker runs under unsupported local Node #then it bootstraps through the launcher [1.98ms]
+
+test/session-start-node-support.test.ts:
+(pass) CodeGraph SessionStart worker Node support > #given an unsupported local Node and a PATH CodeGraph command with auto provisioning disabled #when worker runs #then it skips without touching the workspace [1.63ms]
+(pass) CodeGraph SessionStart worker Node support > #given an unsupported local Node and a PATH CodeGraph command #when auto provisioning succeeds #then it bootstraps with the provisioned binary [3.15ms]
+(pass) CodeGraph SessionStart worker Node support > #given an unsupported local Node but bundled CodeGraph resolves through CODEGRAPH_NODE_BIN #when worker runs #then it bootstraps with the compatible runtime [0.88ms]
+
+test/serve.test.ts:
+(pass) runCodegraphServe > #given CodeGraph is unresolved #when serving MCP #then exits non-zero with a one-line skip hint [0.11ms]
+(pass) runCodegraphServe > #given CodeGraph resolves #when serving MCP #then execs codegraph serve --mcp with inherited stdio and telemetry disabled [0.97ms]
+(pass) runCodegraphServe > #given an unsupported local Node #when serving MCP #then skips fast without spawning codegraph [0.20ms]
+(pass) runCodegraphServe > #given an unsupported local Node but the unsafe override is set #when serving MCP #then it still spawns codegraph [0.15ms]
+(pass) runCodegraphServe > #given an unsupported local Node but CODEGRAPH_NODE_BIN resolves a bundled shim with Node 22 #when serving MCP #then it spawns the compatible runtime [0.14ms]
+(pass) runCodegraphServe > #given OMO_CODEGRAPH_BIN points at an explicit command #when local Node is unsupported #then serve trusts the configured command [0.21ms]
+(pass) runCodegraphServe > #given OMO_CODEGRAPH_BIN points at a missing path #when serving MCP #then exits before spawn [0.17ms]
+(pass) runCodegraphServe > #given Codex SOT disables CodeGraph #when serving MCP #then exits non-zero with a disabled hint [1.99ms]
+(pass) runCodegraphServe > #given Windows Codex SOT install_dir has codegraph.cmd #when serving MCP #then it resolves there and exports CODEGRAPH_INSTALL_DIR [0.99ms]
+(pass) runCodegraphServe > #given Windows OMO_CODEGRAPH_BIN is a Node script #when resolving serve invocation #then Node executes the script path [0.03ms]
+(pass) runCodegraphServe > #given Windows CodeGraph resolves to a cmd shim #when resolving serve invocation #then cmd.exe executes the shim [0.02ms]
+(pass) runCodegraphServe > #given built serve entry #when invoked with a fake CodeGraph binary #then it runs serve mcp exactly once [205.93ms]
+(pass) runCodegraphServe > #given built cli entry #when invoked with a fake CodeGraph binary #then it runs serve mcp exactly once [182.25ms]
+
+test/serve-provision.test.ts:
+(pass) runCodegraphServe provisioning > #given CodeGraph is unresolved #when serving MCP #then provisions CodeGraph before spawning [0.17ms]
+
+test/hook.test.ts:
+(pass) CodeGraph SessionStart hook > #given hook session-start cli args #when invoked with empty JSON input #then it emits valid JSON and exits zero [2.86ms]
+(pass) CodeGraph SessionStart hook > #given CodeGraph MCP reports an uninitialized project #when PostToolUse fires #then it emits OMO global-store init guidance [0.39ms]
+(pass) CodeGraph SessionStart hook > #given real CodeGraph status output has no MCP path phrase #when PostToolUse fires #then it emits OMO global-store init guidance [0.07ms]
+(pass) CodeGraph SessionStart hook > #given CodeGraph is disabled by Codex SOT config #when SessionStart fires #then it skips without spawning [0.13ms]
+(pass) CodeGraph SessionStart hook > #given HOME OMO config disables Codex CodeGraph #when SessionStart fires #then it skips without spawning [1.61ms]
+(pass) CodeGraph SessionStart hook > #given project Codex SOT disables global CodeGraph enablement #when SessionStart fires #then project config wins [1.48ms]
+(pass) CodeGraph SessionStart hook > #given env disables CodeGraph over SOT enablement #when SessionStart fires #then it skips without spawning [0.88ms]
+(pass) CodeGraph SessionStart hook > #given CodeGraph is enabled #when SessionStart fires #then it detaches a background worker immediately [0.46ms]
+(pass) CodeGraph SessionStart hook > #given CodeGraph cannot be resolved or provisioned #when worker runs #then it logs a graceful skip [1.03ms]
+(pass) CodeGraph SessionStart hook > #given CodeGraph is unavailable and auto provisioning is disabled #when worker runs #then it leaves the project untouched [1.15ms]
+(pass) CodeGraph SessionStart hook > #given Windows install_dir has codegraph.cmd #when worker resolves provisioned CodeGraph #then it uses the cmd shim [1.74ms]
+(pass) CodeGraph SessionStart hook > #given Windows codegraph.cmd #when default worker runner builds invocation #then it runs through cmd.exe [0.03ms]
+(pass) CodeGraph SessionStart hook > #given non-Windows codegraph command #when default worker runner builds invocation #then it executes directly
+(pass) CodeGraph SessionStart hook > #given resolved CodeGraph status #when worker runs #then it runs status before init or sync [2.03ms]
+(pass) CodeGraph SessionStart hook > #given malformed hook input with CodeGraph disabled #when SessionStart fires #then it stays silent and exits zero [0.14ms]
+(pass) CodeGraph SessionStart hook > #given plugin hook config #when inspected #then CodeGraph is registered after bootstrap SessionStart [0.28ms]
+
+test/serve-node-support.test.ts:
+(pass) runCodegraphServe node support > #given Node is too new and no command resolves #when serving MCP #then the unsupported-node hint wins [0.23ms]
+
+ 36 pass
+ 0 fail
+ 102 expect() calls
+Ran 36 tests across 6 files. [442.00ms]
