+++ .omo/evidence/20260620-codegraph-bootstrap-node26/summary.md
@@ -3,6 +3,8 @@
 ## Regression
 
 - `focused-component-checks.txt` records `bun run typecheck`, `bun run build`, and `bun run test` for `packages/omo-codex/plugin/components/codegraph`.
+- `focused-component-checks-after-test-split.txt` records the focused typecheck and component test rerun after moving Node-support worker cases into `test/session-start-node-support.test.ts`.
+- `../pr-5448-code-review.md` and `../pr-5448-security-safety-code-review.md` record explicit programming/remove-ai-slops and security review perspectives.
 - `test-codex.txt` records the full `bun run test:codex` gate. It completed with 354 passing Node tests and no failures.
 - `codex-qa-common-self-check.txt` proves the Codex QA isolation helpers pass.
 - `codex-qa-hook-unit-probe.txt` proves the deterministic hook probe still passes.
